+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added resize mode options
+- Added output size options for emotes (28px, 56px, 112px) and badges (18px, 36px, 72px)
+- Added an image icon to drag area
 - Changed image resizing library to Jimp.
 
 ## [0.5.0] - 2020-08-23